refactor: Recursive and number parser
